# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API prueba previa junior PHP",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/socio": {
            "get": {
                "tags": [
                    "Socio"
                ],
                "summary": "Mostrar socios",
                "operationId": "cea57f265aec9477376b4ec903fa5d52",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los socios.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Socio"
                ],
                "summary": "Crear socio",
                "operationId": "2c03b4a6a4709d6107bf8c6829cacd37",
                "requestBody": {
                    "request": "SocioStoreRequest",
                    "description": "Contiene los datos del socio a añadir",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "socio": {
                                    "nombre": "Prueba",
                                    "apellidos": "Socios",
                                    "dni": "12343212E",
                                    "email": "example@example.com",
                                    "contrasena": "12341234"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Socio creado de manera exitosa",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/socio/{id}": {
            "get": {
                "tags": [
                    "Socio"
                ],
                "summary": "Mostrar un socio",
                "operationId": "a3a4ecd0db35abc43679663bb1345770",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id del socio a mostrar",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar un socio.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Socio"
                ],
                "summary": "Modificar socio",
                "operationId": "8b52016b547052f39a67320f88f8c67b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id del socio a modificar",
                        "required": true
                    }
                ],
                "requestBody": {
                    "request": "SocioUpdateRequest",
                    "description": "Contiene los datos del socio a modificar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "socio": {
                                    "nombre": "PruebaPut",
                                    "conntrasena": "12341234"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Socio modificado de manera exitosa",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Socio"
                ],
                "summary": "Borrar un socio",
                "operationId": "eed823c79231149746b11a64354eaa14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id del socio a eliminar",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Socio eliminado correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}